Sime_energy/sime-backend-main/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.41
pydantic==2.11.7
botocore>=1.35.0
boto3>=1.35.0
amazon-textract-response-parser==1.0.1
alembic==1.13.1
pdf2image
python-multipart
pandas==2.3.1
chardet==5.2.0
openpyxl==3.1.5
mistralai==1.9.2
pytesseract==0.3.10
Pillow==11.3.0
opencv-python==4.10.0.84
email-validator==2.1.0
python-jose[cryptography] 
pydantic-settings 
python-multipart
supabase==2.24.0
dotenv
weasyprint
requests